Parse root structure
The root structure contains to unknown 32-bit fields with the value
of 0x1, followed by a 32-bit field containing the number of WMI
classes which will follow.

Each WMI class is prefixed with a 32-bit field containing the length
of the whole class strucuture (including the length field).

All information come from the parse_root() function inside the
bmfparse tool.

Signed-off-by: Armin Wolf <armin.wolf@mailbox.tu-dresden.de>