Parse WMI properties
It turns out WMI variables and class properties are actually one and the
same, just like inside the MS-WMI specification.

The WMI property starts with a 32-bit field containing its data type.
After that a 32-bit field with an unknown purpose and a constant value
of 0 follows. Then two 32-bit fields follow containing the offset of the
property value and the property qualifiers. Those offsets start after
the field for the qualifiers offset and contain a special value
(0xFFFFFFFF) if the associated data is not present.

Finally a UTF-16-LE string containing the name of the property follows
and (if available) the value of the property and an array containing
qualifiers.

Most of the information comes from the parse_class_variable() and
parse_class_property() functions inside the bmfparse tool.

The remaing details where derived from the encoding of WMI properties
inside WMI object instances, since those contain both a value and
qualifiers (even if the latter is just an empty array).

Signed-off-by: Armin Wolf <armin.wolf@mailbox.tu-dresden.de>