Add support for WMI class instances
It turns out that WMI class instances are encoded the same way as
ordinary WMI classes, the only difference is the value of the type
field (0 for classes, 1 for class instances).

Add support for this to better test variables with values which are
only available when defining WMI class instances.

Signed-off-by: Armin Wolf <armin.wolf@mailbox.tu-dresden.de>