Parse WMI method structure
The WMI method structure is similar the the WMI property structure,
except that the meaning of the "data_type" field seems to be
different (can be either 0x200D or 0x0) and that the "value" part
contains the WMI method parameters.

The necessary information comes from the parse_class_method()
function inside the bmfparse tool.

Signed-off-by: Armin Wolf <armin.wolf@mailbox.tu-dresden.de>