Parse WMI methods
A WMI method is encoded as an property containing an object array
with two entries.

Each object has the name "__PARAMETERS" and contains the WMI method
parameters as its properties. The first object contains the input
parameters while the second object contains the output parameters.
The special output parameter called "ReturnValue" determines the
return type of the WMI method.

The necessary knowledge was taken from the bmfparse tool and the
MS-WMIIO specification.

Signed-off-by: Armin Wolf <armin.wolf@mailbox.tu-dresden.de>