Add support for object alias values
As described in [1] both class and instance definitions can have
an optional alias value that can be used to refer to the associated
class or instance. This alias name is stored inside the __ALIAS
property.

Add support for this reserved property so that __ALIAS does not show
up as a regular property when parsing WMI objects. Also add some MOF
files for testing purposes.

[1] https://learn.microsoft.com/en-us/windows/win32/wmisdk/creating-an-alias

Signed-off-by: Armin Wolf <W_Armin@gmx.de>